Support two special accent-cancel forms
In Telex, the duplication of modifier is used to cancel the
accents/hats. For example, "aaa" and "aww" will print "aa"
and "aw" (resp.). In VNI, they are "a66" and "a88".

The 3rd form of cancel form is sth like "ass" (telex) which
can be tranformed natually into VNI code "a11" (vni)