find_missing will add cancel forms if missing
The cancel form, e.g, "unss" -> "uns" is required. We skip the
this cancel form if there is a hat (^) or horn (w) in the combination.
This is because we support cancel form to help users to type English,
and there is not (^), (w) in English (but possibly in German/French)